feat: some improvements to all script

+++ all.sh
@@ -4,32 +4,13 @@ usage() {
   echo "Usage: $0 [-h][--help][-k FILTER]"
   echo "  -h | --help     Display this help message"
   echo "  -k FILTER       Filter devices by name"
-    echo "Available Devices:"
-  echo "------------------"
-  echo "switchboxd_20190808 (port: 5001)"
-  echo "switchboxd_20200229 (port: 5002)"
-  echo "switchboxd_20200831 (port: 5003)"
-  echo "switchbox_20180604 (port: 5011)"
-  echo "switchbox_20190808 (port: 5012)"
-  echo "switchbox_20200229 (port: 5013)"
-  echo "switchbox_20200831 (port: 5014)"
-  echo "switchbox_20220114 (port: 5015)"
-  echo "floodsensor_20200831 (port: 5021)"
-  echo "floodsensor_20210413 (port: 5022)"
-  echo "windrainsensor_20200831 (port: 5031)"
-  echo "windrainsensor_20210413 (port: 5032)"
-  echo "wlightbox_20200229 (port: 5041)"
-  echo "shutterbox_20190911 (port: 5051 to 5059)"
-  echo "gatebox_20230102 (port: 5061 to 5063)"
-  echo "shutterbox_20190911 (port: 5953)"
-  echo "------------------"
-  exit 1
 }
 
 while getopts ":hk:" opt; do
   case ${opt} in
     h)
       usage
+      HELPMODE=1
       ;;
     k)
       MODULE_FILTER=$OPTARG
@@ -37,19 +18,16 @@ while getopts ":hk:" opt; do
     \?)
       echo "Invalid option: $OPTARG" 1>&2
       usage
+      exit 1
       ;;
     :)
       echo "Option -$OPTARG requires an argument." 1>&2
       usage
+      exit 1
       ;;
   esac
 done
 shift $((OPTIND -1))
-
-if [[ "$1" == "-h" ]] || [[ "$1" == "--help" ]]; then
-  display_available_devices
-  exit 0
-fi
 
 while [[ $# -gt 0 ]]; do
   echo "Too many arguments: ${@}"
@@ -67,46 +45,57 @@ function device() {
   fi
 
   echo "port[$port]: $module"
+  # abort running things in helpmode
+  if [[ ! -z ${HELPMODE} ]]; then
+    return
+  fi
+
   flask --app devices.$module run --port $port $extra | grep -v "This is a development server" & # who cares
 
   if command -v dns-sd &> /dev/null; then
     dns-sd -P $module _bbxsrv local. $port localhost 127.0.0.1 &
   fi
 }
-trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
-# ---
-device switchboxd_20190808 5001 --reload
-device switchboxd_20200229 5002 --reload
-device switchboxd_20200831 5003 --reload
-# ---
-device switchbox_20180604 5011 --reload
-device switchbox_20190808 5012 --reload
-device switchbox_20200229 5013 --reload
-device switchbox_20200831 5014 --reload
-device switchbox_20220114 5015 --reload
-# ---
-device floodsensor_20200831 5021 --reload
-device floodsensor_20210413 5022 --reload
-# ---
+# --- switchbox family
+device switchboxd_20190808     5001 --reload
+device switchboxd_20200229     5002 --reload
+device switchboxd_20200831     5003 --reload
+device switchbox_20180604      5011 --reload
+device switchbox_20190808      5012 --reload
+device switchbox_20200229      5013 --reload
+device switchbox_20200831      5014 --reload
+device switchbox_20220114      5015 --reload
+# --- floodsensor family
+device floodsensor_20200831    5021 --reload
+device floodsensor_20210413    5022 --reload
+# --- windrainsensor family
 device windrainsensor_20200831 5031 --reload
 device windrainsensor_20210413 5032 --reload
+# --- lightbox family
+device wlightbox_20200229      5041 --reload
+# --- multisensor family
+device multisensor_20220114    5051 --reload
+
+# --- complex setups
+MODE=1 VARIANT=segmented            device shutterbox_20190911 5151
+MODE=2 VARIANT=nocalib              device shutterbox_20190911 5152
+MODE=3 VARIANT=tilt                 device shutterbox_20190911 5153
+MODE=4 VARIANT=window               device shutterbox_20190911 5155
+MODE=5 VARIANT=material             device shutterbox_20190911 5156
+MODE=6 VARIANT=awning               device shutterbox_20190911 5157
+MODE=7 VARIANT=screen               device shutterbox_20190911 5158
+MODE=8 VARIANT=curtain              device shutterbox_20190911 5159
 # ---
-device wlightbox_20200229 5041 --reload
-# ---
-MODE=1 VARIANT=segmented            device shutterbox_20190911 5051
-MODE=2 VARIANT=nocalib              device shutterbox_20190911 5052
-MODE=3 VARIANT=tilt                 device shutterbox_20190911 5053
-MODE=4 VARIANT=window               device shutterbox_20190911 5055
-MODE=5 VARIANT=material             device shutterbox_20190911 5056
-MODE=6 VARIANT=awning               device shutterbox_20190911 5057
-MODE=7 VARIANT=screen               device shutterbox_20190911 5058
-MODE=8 VARIANT=curtain              device shutterbox_20190911 5059
-# ---
-MODE=0 VARIANT=step-by-step   device gatebox_20230102 5061
-MODE=1 VARIANT=only-open      device gatebox_20230102 5062
-MODE=2 VARIANT=open-close     device gatebox_20230102 5063
+MODE=0 VARIANT=step-by-step   device gatebox_20230102 5161
+MODE=1 VARIANT=only-open      device gatebox_20230102 5162
+MODE=2 VARIANT=open-close     device gatebox_20230102 5163
+
 # --- faulty devices ---
 FAULTY=1 MODE=3 VARIANT=tilt-faulty device shutterbox_20190911 5953
-while true; do
-  sleep 0.1
-done
+
+if [[ -z "${HELPMODE}" ]]; then
+  trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
+  while true; do
+    sleep 0.1
+  done
+fi
